requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.3
gunicorn==20.0.4
idna==2.9
multitasking==0.0.9
numpy==1.18.1
pandas==1.0.2
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
sqlparse==0.3.0
urllib3==1.25.8
yfinance==0.1.54
